Changed file extension from .js to .ts and resolved a few linter errors
