+++ dfm-sideline-sidekick-app/download/checkDevice/checkDevice.ts
@@ -1,7 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-return */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-
 import * as Device from "expo-device";
 
 // Returns true if device, false if emulator
