--- dfm-sideline-sidekick-app/download/connection/checkConnection.js
@@ -1,13 +1,0 @@
-import NetInfo from '@react-native-community/netinfo';
-
-export const checkConnection = async () => {
-    let connected;
-
-    await NetInfo.fetch().then(state => {
-        console.log('Connection type', state.type);
-        console.log('Is connected?', state.isConnected);
-        connected = state.isConnected;
-    });
-
-    return connected;
-}
