--- dfm-sideline-sidekick-app/download/createResumable/createResumable.js
@@ -1,10 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-import * as FileSystem from 'expo-file-system';
-
-export const createResumable = (url, fileDir, fileName) => {
-    return FileSystem.createDownloadResumable(
-        url,
-        fileDir + fileName,
-        {}
-    )
-}
