+++ dfm-sideline-sidekick-app/download/createResumable/createResumable.ts
@@ -1,0 +1,5 @@
+import * as FileSystem from "expo-file-system";
+
+export const createResumable = (url: string, fileDir: string, fileName: string) => {
+  return FileSystem.createDownloadResumable(url, fileDir + fileName, {});
+};
