--- dfm-sideline-sidekick-app/download/versionControl/getCurrentVersion.js
@@ -1,5 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-export const getCurrentVersion = async (versionUrl) => {
-    const versionResponse = await fetch(versionUrl);
-    return versionResponse.json();
-}
