+++ dfm-sideline-sidekick-app/download/versionControl/getCurrentVersion.ts
@@ -1,0 +1,4 @@
+export const getCurrentVersion = async (versionUrl: string) => {
+  const versionResponse = await fetch(versionUrl);
+  return versionResponse.json();
+};
